backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.9",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest-cov>=7.0.0",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
